Refactored code: moved REST api to specified package
